rename to dashboard in catalogue-solutions
