+++ app/pages/sections/order-items/catalogue-solutions/dashboard/manifest.json
